requirements.txt
altair==6.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
blinker==1.9.0
blis==1.3.3
cachetools==7.0.6
catalogue==2.0.10
certifi==2026.4.22
charset-normalizer==3.4.7
click==8.3.3
cloudpathlib==0.23.0
colorama==0.4.6
confection==1.3.3
contourpy==1.3.3
cycler==0.12.1
cymem==2.0.13
distro==1.9.0
fonttools==4.62.1
gitdb==4.0.12
GitPython==3.1.48
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.9
mdurl==0.1.2
murmurhash==1.0.15
narwhals==2.20.0
numpy==2.4.4
openai==2.33.0
packaging==26.2
pandas==3.0.2
pillow==12.2.0
preshed==3.0.13
protobuf==7.34.1
pyarrow==24.0.0
pydantic==2.13.3
pydantic_core==2.46.3
pydeck==0.9.2
Pygments==2.20.0
PyMuPDF==1.27.2.3
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
referencing==0.37.0
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
smart_open==7.6.0
smmap==5.0.3
sniffio==1.3.1
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.3
streamlit==1.56.0
tenacity==9.1.4
thinc==8.3.13
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typer==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
wasabi==1.1.3
watchdog==6.0.0
weasel==1.0.0
wrapt==2.1.2
